new_makefile_solong

+++ CURSUS/so_long/incl/MLX42/src/utils/mlx_utils.c
@@ -3,10 +3,10 @@
 /*                                                        ::::::::            */
 /*   mlx_utils.c                                        :+:    :+:            */
 /*                                                     +:+                    */
-/*   By: W2Wizard <w2.wizzard@gmail.com>              +#+                     */
+/*   By: W2Wizard <main@w2wizard.dev>                 +#+                     */
 /*                                                   +#+                      */
 /*   Created: 2022/01/03 20:13:17 by W2Wizard      #+#    #+#                 */
-/*   Updated: 2022/06/29 15:34:25 by lde-la-h      ########   odam.nl         */
+/*   Updated: 2022/11/22 10:56:09 by jvan-hal      ########   odam.nl         */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -18,12 +18,12 @@
  * Function to read a file stream line by line, reusing the same output pointer.
  * Since the same output pointer is reused it should only be freed once, either on success or failure.
  * This function is made to be somewhat similar to getline.
- * Getline cant be used directly since its not standard and there for not available on all platforms.
+ * Getline can't be used directly since it's not standard and therefore not available on all platforms.
  *
  * @param out Pointer to store output string.
  * @param out_size Pointer to store output strings length.
  * @param file File stream to read from.
- * @return True if line was read, false if EOF was reached or an error ocurred.
+ * @return True if line was read, false if EOF was reached or an error occurred.
  */
 bool mlx_getline(char** out, size_t* out_size, FILE* file)
 {
@@ -84,7 +84,7 @@ uint64_t mlx_fnv_hash(char* str, size_t len)
  * 
  * @param count The amount of args provided.
  * @param ... Any form of pointer.
- * @return False, this is simply for convinience when necessary.
+ * @return False, this is simply for convenience when necessary.
  */
 bool mlx_freen(int32_t count, ...)
 {
@@ -118,7 +118,7 @@ uint32_t mlx_rgba_to_mono(uint32_t color)
 
 //= Public =//
 
-int32_t mlx_get_time(void)
+double mlx_get_time(void)
 {
 	return (glfwGetTime());
 }
